// src/pagesTypes/helpUkraine.types.ts
export enum HelpUkraineSelectItems {
  SelectHelp = '--Select Help--',
  FinancialAssistance = 'Financial Assistance',
  PersonalInvolvement = 'Personal Involvement'
}

export enum HelpUkraineSelectItemsUk {
  FinancialAssistance = 'Фінансова Допомога',
  PersonalInvolvement = 'Особиста Участь'
}

export enum HelpUkraineSelectId {
  FinancialAssistance = 'financialAssistance',
  PersonalInvolvement = 'personalInvolvement',
}